pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-futures-trading"
version = "0.1.0"
description = "Comprehensive Bitcoin price prediction and automated futures trading system for Binance"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Quantitative Trading Team", email = "trading@example.com"}
]

dependencies = [
    # Data Collection & APIs
    "ccxt>=4.0.0",
    "python-binance>=1.0.17",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    
    # Data Processing & Storage
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "polars>=0.19.0",
    "pyarrow>=13.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "clickhouse-driver>=0.4.0",
    
    # Time Series & Feature Engineering
    "ta-lib>=0.4.28",
    "ta>=0.10.2",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    
    # Machine Learning & Deep Learning
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "transformers>=4.30.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "optuna>=3.0.0",
    
    # Backtesting
    "backtrader>=1.9.76",
    "freqtrade>=2023.10",
    
    # Production & Monitoring
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "prometheus-client>=0.18.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    
    # Utilities
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

data = [
    "glassnode-api>=1.0.0",
    "kaiko-api>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

